moved all functions to utilites
